Change Dockerfile base image to Ubuntu 22.04
Avoids the slow linker issue of Debian 12. mips-linux-gnu-ld just runs too slow. It takes more than a minute to complete while it should take less than a second. The current version of Debian sid would fix this, but I don't want to use unstable.

Not using Ubuntu 24 as there is a compilation problem involving the boot code (aspboot, gspboot, etc.) not being found. The assembler seems to be looking at the /bin directory at the root of the disk rather than at the local bin directory that's at the root of the repo. I don't have the time to investigate this right now, so we'll be using an older Ubuntu version.